requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
pyyaml>=6.0.3
sentence-transformers>=2.6.1
faiss-cpu>=1.8.0

